GeoNode REST API Manager

GeoNode REST API Manager

GeoNode REST API Manager

Format

R6Class object.

Value

Object of R6Class with methods for communication with the REST API of a GeoNode instance.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Public fields

verbose.info

if geonode4R logs have to be printed

verbose.debug

if curl logs have to be printed

loggerType

the type of logger

url

the Base url of GeoNode

Methods

Public methods

GeoNodeManager$logger()

GeoNodeManager$INFO()

GeoNodeManager$WARN()

GeoNodeManager$ERROR()

GeoNodeManager$new()

GeoNodeManager$getUrl()

GeoNodeManager$connect()

GeoNodeManager$getCategories()

GeoNodeManager$getCategory()

GeoNodeManager$getResourceByUUID()

GeoNodeManager$getResource()

GeoNodeManager$deleteResource()

GeoNodeManager$upload()

GeoNodeManager$uploadMetadata()

GeoNodeManager$getDataset()

GeoNodeManager$clone()

Method logger()

Prints a log message

Usage

GeoNodeManager$logger(type, text)

Arguments

type

type of log, "INFO", "WARN", "ERROR"

text

text

Method INFO()

Prints an INFO log message

Usage

GeoNodeManager$INFO(text)

Arguments

text

text

Method WARN()

Prints an WARN log message

Usage

GeoNodeManager$WARN(text)

Arguments

text

text

Method ERROR()

Prints an ERROR log message

Usage

GeoNodeManager$ERROR(text)

Arguments

text

text

Method new()

This method is used to instantiate a GeoNodeManager with the url of the GeoNode and credentials to authenticate (user/pwd).

By default, the logger argument will be set to NULL (no logger). This argument accepts two possible values: INFO: to print only geosapi logs, DEBUG: to print geosapi and CURL logs.

The keyring_backend can be set to use a different backend for storing the GeoNode user password with keyring (Default value is 'env').

Usage

GeoNodeManager$new(url, user, pwd, logger = NULL, keyring_backend = "env")

Arguments

url

url

user

user

pwd

pwd

logger

logger

keyring_backend

keyring backend. Default is 'env'

Method getUrl()

Get URL

Usage

GeoNodeManager$getUrl()

Returns

the Geoserver URL

Method connect()

Connects to geoServer

Usage

GeoNodeManager$connect()

Returns

TRUE if connected, raises an error otherwise

Method getCategories()

Get categories

Usage

GeoNodeManager$getCategories(raw = FALSE)

Arguments

raw

Controls the output. Default will return an object of class data.frame.

Returns

an object of class list

Method getCategory()

Get category

Usage

GeoNodeManager$getCategory(id, raw = FALSE)

Arguments

id

category id

raw

Controls the output. Default will return an object of class data.frame.

Returns

an object of class list

Method getResourceByUUID()

Get resource by UUID

Usage

GeoNodeManager$getResourceByUUID(uuid)

Arguments

uuid

resource uuid (or semantic id if used in place of uuid)

Returns

an object of class list

Method getResource()

Get resource

Usage

GeoNodeManager$getResource(id)

Arguments

id

resource id

Returns

an object of class list

Method deleteResource()

Deletes a resource

Usage

GeoNodeManager$deleteResource(id)

Arguments

id

resource (either a dataset or document) id

Returns

TRUE if deleted, FALSE otherwise

Method upload()

Uploads resource files

Usage

GeoNodeManager$upload(files)

Arguments

files

files

Returns

an object of class list giving the upload status

Method uploadMetadata()

Uploads ISO 19115 dataset metadata

Usage

GeoNodeManager$uploadMetadata(id, file)

Arguments

id

dataset id

file

a metadata XML file following ISO 19115 specification

Returns

an object

Method getDataset()

Get dataset standardized metadata

Usage

GeoNodeManager$getDataset(id)

Arguments

id

dataset id

Returns

an object of class list

Method clone()

The objects of this class are cloneable with this method.

Usage

GeoNodeManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \dontrun{
   GeoNodeManager$new("http://localhost:8080", "user", "password")
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to localhost port 8080 after 0 ms: Connection refused
# }